Potential Change to ProgOps.LoginCommand()
This file shows a change to the LoginCommand function that changes the return type from void to char. The char this function returns is then used to determine which course of action the program will take. (Login to Admin, Employee, or customer, report bad credentials, and report a database error.)