Add files via upload
Added btnLogin for the Password Recovery Form so the user can return to the login screen and then exit from there but they can not exit via this form. Altered text on btnRecover to read “Send Email”.
Added email format validation on frmPasswordRecovery with a ToolTip (But it’s not working correctly.